Add check for existing fullscreen
The script now attempts to get the screen size and window size,
and checks if the window is fullscreened or not already.

If the window is fullscreened, it skips sending the F11 fullscreen key.

Regardless of browser window size, the 'f' key is still sent for
YouTube videos -- as no current use cases would imply an already
fullscreened YouTube video.

Also added '[COMMAND]' output for each command that gets ran.